+++ modules/client/CHANGELOG.md
@@ -19,6 +19,11 @@ TEMPLATE:
 
 ## [UPCOMING]
 
+- Add support for v1.0.0 and v1.3.0 build metadata
+- Fixed toLowerCase in subgraphQuery
+
+## [1.20.1]
+
 - Fixed pluginAddress comparison in `validatePluginUpdateProposal`
 
 ## [1.20.0]